Não dá pra pular a cena de vitória tão rápido
- Adicionado 1s de espera antes que possa pular a cena de vitória.

+++ Assets/Resources/Scripts/SceneLoader.cs
@@ -9,10 +9,17 @@ public class SceneLoader : MonoBehaviour {
     private bool upperReady = false, bottomReady = false;
     public Image upperPlay, bottomPlay;
     public GameObject upperBlock, bottomBlock;
+    private float timePassed = 0;
+
+    private void Update() {
+        if (timePassed < 1f) timePassed += Time.deltaTime;
+    }
 
     public void ChangeScene(string s) {
-        PlayerPrefs.DeleteAll();
-        SceneManager.LoadScene(s);
+        if (timePassed >= 1f) {
+            PlayerPrefs.DeleteAll();
+            SceneManager.LoadScene(s);
+        }
 	}
 
     public void SetUpperReady() {
@@ -21,8 +28,11 @@ public class SceneLoader : MonoBehaviour {
             upperPlay.color = ColorManager.colors[(int)ColorManager.upperPlayerColor];
         else
             upperPlay.color = Color.white;
-        if (upperReady && bottomReady) ChangeScene("Gameplay");
-        upperBlock.SetActive(upperReady);
+        if (upperReady && bottomReady) {
+            PlayerPrefs.DeleteAll();
+            SceneManager.LoadScene("Gameplay");
+        }
+            upperBlock.SetActive(upperReady);
     }
 
     public void SetBottomReady() {
@@ -31,7 +41,10 @@ public class SceneLoader : MonoBehaviour {
             bottomPlay.color = ColorManager.colors[(int)ColorManager.bottomPlayerColor];
         else
             bottomPlay.color = Color.white;
-        if (upperReady && bottomReady) ChangeScene("Gameplay");
+        if (upperReady && bottomReady) {
+            PlayerPrefs.DeleteAll();
+            SceneManager.LoadScene("Gameplay");
+        }
         bottomBlock.SetActive(bottomReady);
     }
